CHROMIUM: low_mem: make ram_vs_swap_weight configurable
As a temporary fix to OOM kill deadlocks, we want to try to
avoid OOMs or at least reduce the number of occurrences.
So we make the ram_vs_weight parameter run-time configurable,
with two possible uses in mind: 1. leave as is (the default)
and 2. on specific devices, set it to infinity so that
available swap is completely ignored in the available memory
calculation.

This also adds the "available" sysfs entry, so that chrome
doesn't have to duplicate the available memory calculations.

BUG=chromium:727707
TEST=played around with sysfs

Change-Id: I9f9391c07227200299f6bb570687dc883feca6cb
Reviewed-on: https://chromium-review.googlesource.com/522822
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>

[rebase419(groeck): Squashed commit:
	FIXUP: CHROMIUM: low_mem: make ram_vs_swap_weight configurable]

Change-Id: I5319592f1b9fc04ae8278885e0e231dacdabc5da

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>