arm64: debug: Separate debug hooks based on target exception level
Mixing kernel and user debug hooks together is highly error-prone as it
relies on all of the hooks to figure out whether the exception came from
kernel or user, and then to act accordingly.

Make our debug hook code a little more robust by maintaining separate
hook lists for user and kernel, with separate registration functions
to force callers to be explicit about the exception levels that they
care about.

Reviewed-by: Mark Rutland <mark.rutland@arm.com>
[@kdrag0n: Ported to android-4.19 with adaptations for KASAN hook]

Change-Id: I830ad3dba6e193b6d569f981121f38f1a80e0e73
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>