UPSTREAM: net: ipip: implement header_ops->parse_protocol for AF_PACKET
Ipip uses skb->protocol to determine packet type, and bails out if it's
not set. For AF_PACKET injection, we need to support its call chain of:

    packet_sendmsg -> packet_snd -> packet_parse_headers ->
      dev_parse_header_protocol -> parse_protocol

Without a valid parse_protocol, this returns zero, and ipip rejects the
skb. So, this wires up the ip_tunnel handler for layer 3 packets for
that case.

Acked-by: Willem de Bruijn <willemb@google.com>
Change-Id: I61b74dbd5e3322ce7028b112dcb6b350f7fa7c94
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>