exfat: remove 'rwoffset' in exfat_inode_info
Remove 'rwoffset' in exfat_inode_info and replace it with the parameter of
exfat_readdir().
Since rwoffset is referenced only by exfat_readdir(), it is not necessary
a exfat_inode_info's member.
Also, change cpos to point to the next of entry-set, and return the index
of dir-entry via dir_entry->entry.

Acked-by: Sungjong Seo <sj1557.seo@samsung.com>
Change-Id: I4ba653858731e7a190d9d77f3b1f97563016cddc
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>