BACKPORT: exfat: fix erroneous discard when clear cluster bit
If mounted with discard option, exFAT issues discard command when clear
cluster bit to remove file. But the input parameter of cluster-to-sector
calculation is abnormally added by reserved cluster size which is 2,
leading to discard unrelated sectors included in target+2 cluster.
With fixing this, remove the wrong comments in set/clear/find bitmap
functions.

Fixes: 1e49a94cf707 ("exfat: add bitmap operations")
Cc: stable@vger.kernel.org # v5.7+
Acked-by: Sungjong Seo <sj1557.seo@samsung.com>
[ @ Tashar02: Backport to k4.19 ]

Change-Id: Ie8f28ffaa1357b89fdbeed31fefb4221d8a61fa1

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>