qos: Replace cpumasks with atomic bitwise operations
This is expensive and unneeded, especially since the qos functions are
hot code paths. Mose of the cpumask functions use the bitmap API, which
is also more expensive than just doing some simple operations on a word.

Since we're operating with a CPU count that can fit within a word,
replace the expensive cpumask operations with raw bitwise operations
wherever possible to make the pm_qos framework more efficient.

Co-authored-by: Dušan Uverić <dusan.uveric9@gmail.com>
Change-Id: I842dcb46980a3bb5b1f9449f950d5d494e4d57c9

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>