arm64: lib: Use modern annotations for assembly functions
In an effort to clarify and simplify the annotation of assembly functions
in the kernel new macros have been introduced. These replace ENTRY and
ENDPROC and also add a new annotation for static functions which previously
had no ENTRY equivalent. Update the annotations in the library code to the
new macros.

Change-Id: I879ffc14c2c7312ffd76debd15d3722942c5460b
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>