UPSTREAM: wireguard: main: rename 'mod_init' & 'mod_exit' functions to be module-specific
Rename module_init & module_exit functions that are named
"mod_init" and "mod_exit" so that they are unique in both the
System.map file and in initcall_debug output instead of showing
up as almost anonymous "mod_init".

This is helpful for debugging and in determining how long certain
module_init calls take to execute.

Change-Id: I845caec39827482bf2f3909a412702dbe82b01ba
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>