{X00TD, X01BD}_defconfig: Disable sysrq keys over serial
f the UART console is enabled in the bootloader, this could cause false
triggers, especially if the other side's UART interface has a loose TX
pin (which I've encountered).

Note that sysrq is still enabled because Android init uses it when
shutting down, and it doesn't hurt to keep it enabled for debugging.

Change-Id: Ibf28825417af99da46e8d79317808d33b7fd3c4a
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>