{X00TD, X01BD}_defconfig: Don't randomize module region over a 4 GB range
Randomizing a module is for security purpose to deter exploit
attempts relying on knowledge of the location of kernel internals.

But randomizing the location of module region inside a 4 GB window
to prevent core kernel data leakage is expensive. We are better off
with a limited range. Randomizing module region is expensive anyway.

Change-Id: I9a0029bae04fef6b826d93d64ce0d8f3aa43e070
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>