PM / suspend: Clear wakeups before running PM callbacks
PM callbacks can be used as an indicator that the system is suspending, and
as such, a wakeup may be issued outside the notifier in order to cancel an
ongoing suspend. Clearing wakeups after running the PM callbacks can thus
cause wakeups to be missed. Fix it by simply clearing wakeups prior to
running PM callbacks.

Change-Id: I6b4d3fb4345cbb6345cdb8d0aa1ed6d053173605
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>