BACKPORT: random: remove rng_has_arch_random()
With arch randomness being used by every distro and enabled in
defconfigs, the distinction between rng_has_arch_random() and
rng_is_initialized() is now rather small. In fact, the places where they
differ are now places where paranoid users and system builders really
don't want arch randomness to be used, in which case we should respect
that choice, or places where arch randomness is known to be broken, in
which case that choice is all the more important. So this commit just
removes the function and its one user.

Reviewed-by: Petr Mladek <pmladek@suse.com> # for vsprintf.c
[ Tashar02: Backport to k4.19 ]

Change-Id: I63a9e73ca51112cc9ca545f7ca01df7c1502f341

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>